update max waiting time to uint64_t

+++ lib/fles_libfabric/InputChannelConnection.hpp
@@ -107,7 +107,7 @@ public:
 
     void add_sent_timestamps(uint64_t timestamp) { sent_timestamps_list_.push_back(timestamp); }
 
-    double max_avg=0,max_max=0;
+    uint64_t max_avg=0,max_max=0;
 
 private:
     /// Post a receive work request (WR) to the receive queue
